pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.25.17",
    "dotenv>=0.9.9",
    "langchain>=1.2.15",
    "langchain-openai>=1.2.1",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "flask>=3.0.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
